Remove `RemainingBatchesLen` method
Based on feedback received from @QuentinI: https://github.com/EspressoSystems/optimism-espresso-integration/pull/230#discussion_r2345037174

The only reason `RemainingBatchesLen` exists is to serve as a check, and issue a
warning when things are running.  Even though this matches the existing behavior this
log seems to overlap with the logs corresponding to Undecided Batches which already
log warnings or errors.  As a result this method, and the log that utilizes it, seem to
be unnecessary and should be removed to eliminate noise.

This change removes `RemainingBatchesLen` and the uses of it