Add `RefreshSafeL1Origin` to `EspressoStreamer` interface
Based on feedback provided by @QuentinI: https://github.com/EspressoSystems/optimism-espresso-integration/pull/230#discussion_r2345442127

Since the `RefreshSafeL1Origin` method can potentially be utilized in some places that
do not require a full `Refresh`, and for convenience, it makes sense to allow it to be a
separately exposed method distinct from `Refresh`.

This change adds `RefreshSafeL1Origin` as a required method in the
`EspressoStreamer` interface.