Fix refreshSafeL1Origin logic for Buffered Streamer
The buffered streamer is resetting its read position far more than it needs to, ultimately
reproducing the same issue that was already occurring with the unbuffered version.

In inspecting the behavior with a debugger, it seems we're resetting the reset position
unnecessarily when we receive the same safeL1Origin again.  Additionally, the logic
for determining the read position when the safeL1Origin advances also seems flawed,
in that it is very likely to reset too far in the past.  We really want to keep our relative
read position unless we're explicitly told to Reset.

This change addresses these issues in order to try and smooth out the batches being
returned, and avoid unnecessary reprocessing of previous batches.