Add a Buffered Streamer around Espresso Streamer for batcher (#230)
* Add a Buffered Streamer around Espresso Streamer for batcher

Because the `EspressoStreamer` is getting `Reset` during the Batcher process when
building a batch to submit to the `L1`, it hinders progress of the chain in a reasonable
amount of time, which ultimately causes it to stop creating non-empty blocks.

There are a number of factors that are contributing to this issue, but ultimately the
`Reset` is causing the `EspressoStreamer` to restart from `0` and it takes too long
before it catches back up to the next expected batch.

To remedy this, a Buffer can be used to mitigate this `Reset` and revert to a point
that is much closer to the desired next batch.

In testing it has been observed that the `SafeL2` can sometimes move backwards.  To
safe guard against this, it is better to `Reset` to the `FinalizedL2` position instead of
the `SafeL2` as this behavior has not been observed there.

* Rename EspressoStreamer and EspressoStreamerIFace

Based on feedback provided by @QuentinI in PR review:
https://github.com/EspressoSystems/optimism-espresso-integration/pull/230#discussion_r2345028112

The name `EspressoStreamerIFace` is quite a long name, and the `IFace` suffix isn't
necessary since one could tell it's an `interface` by inspection, or using an `LSP`. The
feedback provided by @QuentinI suggested to rename `EspressoStreamerIFace` to
just `EspressoStreamer` so that it falls in line with our other code approaches.

This change renames `EspressoStreamer` to `BatchStreamer`

This change renames `EspressoStreamerIFace` to `EspressoStreamer`.

* Remove `RemainingBatchesLen` method

Based on feedback received from @QuentinI: https://github.com/EspressoSystems/optimism-espresso-integration/pull/230#discussion_r2345037174

The only reason `RemainingBatchesLen` exists is to serve as a check, and issue a
warning when things are running.  Even though this matches the existing behavior this
log seems to overlap with the logs corresponding to Undecided Batches which already
log warnings or errors.  As a result this method, and the log that utilizes it, seem to
be unnecessary and should be removed to eliminate noise.

This change removes `RemainingBatchesLen` and the uses of it

* Add `RefreshSafeL1Origin` to `EspressoStreamer` interface

Based on feedback provided by @QuentinI: https://github.com/EspressoSystems/optimism-espresso-integration/pull/230#discussion_r2345442127

Since the `RefreshSafeL1Origin` method can potentially be utilized in some places that
do not require a full `Refresh`, and for convenience, it makes sense to allow it to be a
separately exposed method distinct from `Refresh`.

This change adds `RefreshSafeL1Origin` as a required method in the
`EspressoStreamer` interface.

* Fix missed renamed references

* Rename enclave smoke test

* Fix refreshSafeL1Origin logic for Buffered Streamer

The buffered streamer is resetting its read position far more than it needs to, ultimately
reproducing the same issue that was already occurring with the unbuffered version.

In inspecting the behavior with a debugger, it seems we're resetting the reset position
unnecessarily when we receive the same safeL1Origin again.  Additionally, the logic
for determining the read position when the safeL1Origin advances also seems flawed,
in that it is very likely to reset too far in the past.  We really want to keep our relative
read position unless we're explicitly told to Reset.

This change addresses these issues in order to try and smooth out the batches being
returned, and avoid unnecessary reprocessing of previous batches.

* Add Unit Tests for BufferedEspressoStreamer

Fix BufferedEspressoStreamer behavior

While adding unit tests for the BufferedEspressoStreamer it was noticed that the
position of the L2 and the L1 for the Buffer were being mixed together at times.  This
would ultimately lead to very difficult to detect bugs based on observed behavior
alone.

With the addition of the unit tests identifying the issue, the buffer adjustment
behavior has been adjusted to apply to the L2 position in isolation away from the
L1 positions.  The L1 positions will cause a larger Reset in the underlying logic.