Replace gameCtx timeout with shorter value and retries
In the `waitForGameToBePublished` function a `gameCtx` is created with a timeout of
`5 minutes`.  This `gameCtx` is passed to `wait.ForGamePublished` which
immediately adds its own timeout of `2 minutes` which will trigger before the
`5 minute` timeout.  To get around this, we opt for a shorter timeout of `1 minute`,
and we retry up to 5 times to account for the total `5 minute` duration while also
minimizing the total time needed to wait.