Fix `TestSmoke` failing on CI/CD (#237)
The sigp/lighthouse docker image was upgraded from version `v7.1.0` to
`v8.0.0-rc.0` on `2025-09-29`.  Since the image isn't anchored to a version, this
update gets pulled in, and it seems to have breaking changes with our previous
setup.

This change sets the version of the docker image used specifically to `v7.1.0` so
that the previous behavior we're used to is seen.

Additionally, when `TestSmoke` is running, it initially **MUST** download the images
for the docker containers that wer not built in the `Build Devnet` job.  This delays
the launch and running of the DevNet by quite a bit.

Fix this delay by adding a `docker compose pull` setp to `Build Devnet`